+++ src/index.ts
@@ -48,6 +48,11 @@ type WithGoogleSecretsOptions = {
    * The current next config that will be extended
    */
   nextConfig: object;
+
+  /**
+   * Determs if the google secrets should be loaded or not (default = true)
+   */
+  enabled?: boolean;
 };
 
 type Config = {
@@ -114,7 +119,12 @@ async function iterateSecrets(
  * @returns The updated next config
  */
 const withGoogleSecrets = async (options: WithGoogleSecretsOptions) => {
-  const { projectName, filter, filterFn, mapping, versions = {}, nextConfig = {} } = options;
+  const { projectName, filter, filterFn, mapping, versions = {}, nextConfig = {}, enabled = true } = options;
+
+  if (!enabled) {
+    return nextConfig;
+  }
+
   const newNextConfig = { ...nextConfig };
   const secretmanagerClient = new SecretManagerServiceClient();